requirements.txt
fastapi
uvicorn[standard]
pandas
numpy<2.0
scikit-learn
SQLAlchemy
PyMySQL
python-dotenv
tensorflow
keras-ocr
opencv-python-headless
python-multipart
requests
strawberry-graphql
